Implementing grobid-quantities integration via grobid-quantities python client
