+++ .gitignore
@@ -88,3 +88,5 @@ marve/classes.pyc
 marve/grobid.pyc
 marve/measurement.log
 
+**/sample_output
+
