+++ requirements.txt
@@ -1,3 +1,4 @@
 networkx
 pycorenlp
 matplotlib
+grobid-quantities-client
